Backend/requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.2.4
djangorestframework==3.16.0
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.5
pandas==2.3.0
pillow==10.4.0
platformdirs==4.3.6
pyarrow==17.0.0
pygame==2.5.2
pylance==0.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.1
tzdata==2025.2
virtualenv==20.26.5
Werkzeug==3.0.4
